Refactor copy functions and implement local mode
