+++ src/docs-generator/repo-copier.mjs
@@ -1,15 +1,13 @@
-import { exec, execSync } from "node:child_process";
-import { cp } from "node:fs/promises";
 import { rmSync } from "node:fs";
-import * as path from "node:path";
-
-import { promisify } from "node:util";
-
-// Make exec return promises
-const execPromise = promisify(exec);
+import {
+    ghVersion,
+    getLatestRelease,
+    cloneMinimalRepo,
+    copyRepoSrc
+} from "./copy-utils.mjs";
 
 // Organization name
-const orgName = "IFCjs"
+const orgName = "IFCjs";
 
 // List of repositories to include in the docs
 let repositories = [
@@ -18,131 +16,101 @@ let repositories = [
     { name: "fragment", release: "", fallbackBranch: "" },
     { name: "web-ifc", release: "", fallbackBranch: "" },
 ];
-
-// Relative path to store the repos
 const tempDirName = "temp";
-const fullRepoDirName = `${tempDirName}/_repos`;
 
-// Test that the `gh` command-line utility is installed
-const ghVersion = execSync("gh --version", err => {
-    if (err) {
-        console.error("`gh` command is not accessible. Aborting...");
+// Handle command-line arguments
+const validArgs = [ '--local', '--remote' ];
+const [localMode, remoteMode] = validArgs;
 
-        process.exit(1);
-    }
-});
+if (process.argv.length !== 3 ||
+    !process.argv.some((arg) => validArgs.includes(arg))) {
 
-console.info(`Using ${ghVersion}`);
+    console.error(`Please specify any of: ${validArgs}. Exiting...`);
+    process.exit(1);
+}
 
-/* Fetch latest release version. If it errors or if
-   there are no releases yet, return an empty string */
-const getLatestRelease = async (repoName) => {
-    let release = await execPromise(
-        `gh api repos/${orgName}/${repoName}/releases/latest || true`);
-    release = release.stdout;
+// Run "local" mode, using repositories from parent directory
+if (process.argv.includes(localMode)) {
+    const localReposPath = '../';
+    console.info("Using local mode.");
+    // Copy src/ directories inside each repository
+    console.info(`Copying directories from "${localReposPath}"...`);
 
-    try {
-        release = JSON.parse(release).tag_name;
-        if (release === undefined) {
-            release = "";
-        }
-    } catch (e) {
-        if (e instanceof SyntaxError) {
-            release = "";
-        } else {
-            console.error(`Unexpected error while fetching releases
-                           for "${orgName}/${repoName}". Aborting...`);
+    const srcCopies = await Promise.all(
+        repositories.map(async (repo) => {
+            try {
+                const {originalRepoSrcPath, repoSrcCopyPath} = await copyRepoSrc(
+                    repo.name, localReposPath, tempDirName);
 
-            process.exit(2);
-        }
-    }
+                return `'${originalRepoSrcPath}' -> '${repoSrcCopyPath}'`;
+            } catch {
+                return '';
+            }
+        })
+    );
 
-    return release;
-};
+    console.info("Copied 'src/' directories");
+    console.info(srcCopies);
+    
+}
+// Run "remote" mode, copying repositories remotely from GitHub
+else if (process.argv.includes(remoteMode)) {
+    // Relative path to store the repos
+    const fullRepoDirName = `${tempDirName}/_repos`;
+    console.info(`Using remote mode. ${ghVersion()}`);
 
-/* Clone git repository with minimal depth for an specific tag.
-   If no tag or fallback branch specified, it will use the default branch */
-const cloneMinimalRepo = async(repo) => {
-    const repoDirName = path.resolve(`${fullRepoDirName}/${repo.name}`);
+    // Fetch and fill the releases field for each repo
+    console.info("Fetching releases...");
 
-    let command = `gh repo clone ${orgName}/${repo.name} ${repoDirName} -- --depth 1`;
-    let tagName = "(default branch)";
+    repositories = await Promise.all(
+        repositories.map(async (repo) => {
+            return {
+                ...repo,
+                release: await getLatestRelease(orgName, repo.name),
+            };
+        })
+    );
 
-    if (repo.release && repo.release !== "") {
-        command = `${command} -b ${repo.release}`;
-        tagName = repo.release;
-    } else if (repo.fallbackBranch && repo.fallbackBranch !== "") {
-        command = `${command} -b ${repo.fallbackBranch}`;
-        tagName = repo.fallbackBranch;
-    }
+    console.info("Fetched releases:");
+    console.info(repositories);
 
-    await execPromise(`${command}`);
+    // Clone minimal repositories and display the active branch
+    console.info("Cloning minified repositories...");
 
-    return { repoDirName, tagName };
-};
+    const clonedBranches = await Promise.all(
+        repositories.map(async (repo) => {
+            const { repoDirName, tagName } = await cloneMinimalRepo(
+                orgName, repo, fullRepoDirName);
 
-/* Copy the src/ directory inside the repo to the
- * tempDirName directory with the name of the repository */
-const copyRepoSrc = async(repoName) => {
-    const originalRepoSrcPath = path.resolve(
-        `./${fullRepoDirName}/${repoName}/src`);
-    const repoSrcCopyPath = path.resolve(
-        `./${tempDirName}/${repoName}`);
+            return {
+                directory: repoDirName,
+                branch: tagName,
+            };
+        })
+    );
 
-    await cp(originalRepoSrcPath, repoSrcCopyPath, { recursive: true });
+    console.info("Cloned repositories:");
+    console.info(clonedBranches);
 
-    return { originalRepoSrcPath, repoSrcCopyPath };
-};
+    // Copy src/ directories inside each repository
+    console.info("Copying 'src/' directories...");
 
-// Fetch and fill the releases field for each repo
-console.info("Fetching releases...");
+    const srcCopies = await Promise.all(
+        repositories.map(async (repo) => {
+            const {originalRepoSrcPath, repoSrcCopyPath} = await copyRepoSrc(
+                repo.name, fullRepoDirName, tempDirName);
 
-repositories = await Promise.all(
-    repositories.map(async (repo) => {
-        return {
-            ...repo,
-            release: await getLatestRelease(repo.name),
-        };
-    })
-);
+            return `'${originalRepoSrcPath}' -> '${repoSrcCopyPath}'`;
+        })
+    );
 
-console.info("Fetched releases:");
-console.info(repositories);
+    console.info("Copied 'src/' directories");
+    console.info(srcCopies);
 
-// Clone minimal repositories and display the active branch
-console.info("Cloning minified repositories...");
+    // Delete temp/_repo directory since it's not needed anymore
+    console.info(`Removing ${fullRepoDirName} directory...`);
+    rmSync(fullRepoDirName, { recursive: true });
+}
 
-const clonedBranches = await Promise.all(
-    repositories.map(async (repo) => {
-        const { repoDirName, tagName } = await cloneMinimalRepo(repo);
-
-        return {
-            directory: repoDirName,
-            branch: tagName,
-        };
-    })
-);
-
-console.info("Cloned repositories:");
-console.info(clonedBranches);
-
-// Copy src/ directories inside each repository
-console.info("Copying 'src/' directories...");
-
-const srcCopies = await Promise.all(
-    repositories.map(async (repo) => {
-        const {originalRepoSrcPath, repoSrcCopyPath} = await copyRepoSrc(
-            repo.name);
-
-        return `'${originalRepoSrcPath}' -> '${repoSrcCopyPath}'`;
-    })
-);
-
-console.info("Copied 'src/' directories");
-console.info(srcCopies);
-
-// Delete temp/_repo directory since it's not needed anymore
-console.info(`Removing ${fullRepoDirName} directory...`);
-rmSync(fullRepoDirName, { recursive: true });
 
 console.info("Finished copying and setting up repositories!");
